separate domain models for email and password
