+++ src/model/login_user.rs
@@ -1,0 +1,2 @@
+#[derive(Debug)]
+struct LoginUserRequest {}
