+++ src/model/mod.rs
@@ -2,6 +2,10 @@ pub mod api_error;
 pub mod api_response;
 pub mod auth_middleware;
 pub mod login_response;
+pub mod login_user;
 pub mod register_user;
 pub mod token;
 pub mod user;
+pub mod user_email;
+pub mod user_password;
+
